# relay/.assets/swagger.yaml
openapi: 3.1.0
info:
  title: Kamune Relay server
  description: Relay implementation for the Kamune project
  version: 0.1.0
paths:
...
  /queues:
    post:
      operationId: pushToQueue
      tags:
        - queue
      summary: Push a message to the queue
      description: >
        Enqueues a message for a specific sender/receiver/session tuple.
        The message can later be retrieved via `GET /queues`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sender:
                  type: string
                  description: Base64 raw URL encoded public key of the sender
                receiver:
                  type: string
                  description: Base64 raw URL encoded public key of the receiver
                session_id:
                  type: string
                  description: Application-level session identifier
                data:
                  type: string
                  description: Base64 raw URL encoded message payload
              required:
                - sender
                - receiver
                - session_id
                - data
      responses:
        "201":
          description: Message successfully enqueued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                required:
                  - status
                title: PushToQueueCreated
        "400":
          description: Bad request (missing or invalid fields).
        "409":
          description: Queue is full for this sender/receiver/session tuple.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                title: PushToQueueConflict
        "413":
          description: Message payload exceeds the configured maximum size.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                title: PushToQueueTooLarge
    get:
      operationId: popFromQueue
      tags:
        - queue
      summary: Pop a message from the queue
      description: >
        Dequeues the oldest message for a specific sender/receiver/session
        tuple. Messages are returned in FIFO order.
      parameters:
        - in: query
          name: sender
          required: true
          schema:
            type: string
          description: Base64 raw URL encoded public key of the sender
        - in: query
          name: receiver
          required: true
          schema:
            type: string
          description: Base64 raw URL encoded public key of the receiver
        - in: query
          name: session
          required: true
          schema:
            type: string
          description: Application-level session identifier
      responses:
        "200":
          description: A message was successfully dequeued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    description: Base64 raw URL encoded message payload
                required:
                  - data
                title: PopFromQueueOk
        "204":
          description: Queue is empty; no messages available.
        "400":
          description: Bad request (missing or invalid query parameters).
        "404":
          description: No queue found for the given parameters.
  /queues/length:
    get:
      operationId: queueLength
      tags:
        - queue
      summary: Peek at queue depth
      description: >
        Returns the number of pending messages in the queue for a given
        sender/receiver/session tuple without consuming any messages.
      parameters:
        - in: query
          name: sender
          required: true
          schema:
            type: string
          description: Base64 raw URL encoded public key of the sender
        - in: query
          name: receiver
          required: true
          schema:
            type: string
          description: Base64 raw URL encoded public key of the receiver
        - in: query
          name: session
          required: true
          schema:
            type: string
          description: Application-level session identifier
      responses:
        "200":
          description: Queue length returned successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  length:
                    type: integer
                    format: uint64
                    description: Number of pending messages in the queue
                required:
                  - length
                title: QueueLengthOk
        "400":
          description: Bad request (missing or invalid query parameters).
servers:
  - url: http://127.0.0.1:8888
    description: Development server
...